Create theSeventhHouse.py
Matters of the seventh house are always set in the outside world and relationships we tend to create. Its Latin motto is Uxor and this translates to “spouse” with its specific role as a mirror to our own Self found in other people. Although it is often referred to as the house of marriage, it rules all of our relationships, partnerships and things we recognize in other people as a reflection of us. It correlates to the sign of Libra representing our primal balance and our ability to connect with others in the closest possible way. This incarnation provides us with a task to find peace in relation to something specific, and this is seen through the sign on the seventh house cusp, planets inside this house, and its ruler.

P.S. : The astrological information is taken from the "https://www.astrology-zodiac-signs.com/astrology/houses/seventh/" site.